Moved tests to common file
