--- tests/test_data_utils.py
@@ -1,63 +1,0 @@
-import pytest
-import xarray as xr
-
-from ocean_emulators.utils.data import rename_vars
-
-
-def test_rename_vars():
-    """Test renaming variables from OM4 format to standard format."""
-    # Create test dataset with OM4 format variables
-    test_data = {
-        "so_lev_1050_0": (["time", "lat", "lon"], [[[1.0]]]),
-        "thetao_lev_2_5": (["time", "lat", "lon"], [[[2.0]]]),
-        "vo_lev_10_0": (["time", "lat", "lon"], [[[3.0]]]),
-        "zos": (["time", "lat", "lon"], [[[4.0]]]),  # Should remain unchanged
-    }
-    ds = xr.Dataset(
-        test_data,
-        coords={
-            "time": [0],
-            "lat": [0],
-            "lon": [0],
-        },
-    )
-
-    # Apply rename_vars
-    renamed_ds = rename_vars(ds)
-
-    # Test that variables are renamed correctly
-    assert "so_11" in renamed_ds.variables  # 1040.0 is at index 11 in DEPTH_LEVELS
-    assert "thetao_0" in renamed_ds.variables  # 2.5 is at index 0 in DEPTH_LEVELS
-    assert "vo_1" in renamed_ds.variables  # 10.0 is at index 1 in DEPTH_LEVELS
-    assert "zos" in renamed_ds.variables  # Should remain unchanged
-
-    # Test that data values are preserved
-    assert renamed_ds["so_11"].values[0, 0, 0] == 1.0
-    assert renamed_ds["thetao_0"].values[0, 0, 0] == 2.0
-    assert renamed_ds["vo_1"].values[0, 0, 0] == 3.0
-    assert renamed_ds["zos"].values[0, 0, 0] == 4.0
-
-    # Test that original dataset is not modified
-    assert "so_lev_1050_0" in ds.variables
-    assert "thetao_lev_2_5" in ds.variables
-    assert "vo_lev_10_0" in ds.variables
-
-
-def test_rename_vars_invalid_depth():
-    """Test that invalid depth levels raise an error."""
-    # Create test dataset with invalid depth level
-    test_data = {
-        "so_lev_9999_0": (["time", "lat", "lon"], [[[1.0]]]),  # Invalid depth
-    }
-    ds = xr.Dataset(
-        test_data,
-        coords={
-            "time": [0],
-            "lat": [0],
-            "lon": [0],
-        },
-    )
-
-    # Should raise ValueError because 9999.0 is not in DEPTH_LEVELS
-    with pytest.raises(ValueError):
-        rename_vars(ds)
